backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
pytest
httpx
google-generativeai
pyjwt[crypto]
